Edit the README, add one simple example.
The example for python-abp is pretty basic and also quite rough at this
point. It seems that maybe deployment section described in the README
will actually be partially boxes and partially tasks (just like the
publishing task here). I'm not changing the text yet, it will be more
obvious what works when the other examples are fleshed out.